+++ packages/sources/ada-balance/test/integration/adapter.test.ts
@@ -1,26 +1,56 @@
 import { AdapterRequest, Execute } from '@chainlink/types'
 import * as adaBalance from '../../src/'
 import '@cardano-ogmios/client'
+import '../../src/endpoint/ogmios'
+import { InteractionContext } from '@cardano-ogmios/client'
+import 'ws'
+
+const TEST_HTTP_ENDPOINT = 'http://test-ogmios-url.com'
+const TEST_WS_ENDPOINT = 'wss://test-ogmios-url.com'
+
+jest.mock('../../src/endpoint/ogmios', () => ({
+  createInteractionContext: (): InteractionContext => {
+    return {
+      connection: {
+        host: 'test-endpoint',
+        port: 1337,
+        maxPayload: 1000,
+        address: {
+          http: TEST_HTTP_ENDPOINT,
+          webSocket: TEST_WS_ENDPOINT,
+        },
+        tls: false,
+      },
+      socket: jest.mock('ws'),
+      afterEach: jest.fn(),
+    }
+  },
+}))
 
 jest.mock('@cardano-ogmios/client', () => {
   return {
-    utxo: async () => [
-      [
-        {
-          txId: '86a50ff8136e8b5d9f6f249a24330a5b43b657688021980f54bc887bc0cb7f4d',
-          index: 0,
-        },
-        {
-          address:
-            'addr_test1qz87tn9yat3xfutzds43tnj8qw457hk3v46w4028rtnx56v89wjwnrwcvlfm2atvcnnclh3x7thwrl7pgnffaw24mgws0dga4m',
-          value: {
-            coins: 5000000,
-            assets: {},
+    getServerHealth: () => ({
+      lastTipUpdate: '2022-01-15T18:56:42.635812292Z',
+    }),
+    StateQuery: {
+      utxo: async () => [
+        [
+          {
+            txId: '86a50ff8136e8b5d9f6f249a24330a5b43b657688021980f54bc887bc0cb7f4d',
+            index: 0,
           },
-          datum: null,
-        },
+          {
+            address:
+              'addr_test1qz87tn9yat3xfutzds43tnj8qw457hk3v46w4028rtnx56v89wjwnrwcvlfm2atvcnnclh3x7thwrl7pgnffaw24mgws0dga4m',
+            value: {
+              coins: 5000000,
+              assets: {},
+            },
+            datum: null,
+          },
+        ],
       ],
-    ],
+    },
   }
 })
 
@@ -28,7 +58,8 @@ let oldEnv: NodeJS.ProcessEnv
 
 beforeAll(() => {
   oldEnv = JSON.parse(JSON.stringify(process.env))
-  process.env.WS_API_ENDPOINT = 'test-endpoint'
+  process.env.WS_OGMIOS_URL = TEST_WS_ENDPOINT
+  process.env.HTTP_OGMIOS_URL = TEST_HTTP_ENDPOINT
 })
 
 afterAll(() => {
@@ -48,7 +79,10 @@ describe('execute', () => {
       id,
       data: {
         addresses: [
-          'addr_test1qz87tn9yat3xfutzds43tnj8qw457hk3v46w4028rtnx56v89wjwnrwcvlfm2atvcnnclh3x7thwrl7pgnffaw24mgws0dga4m',
+          {
+            address:
+              'addr_test1qz87tn9yat3xfutzds43tnj8qw457hk3v46w4028rtnx56v89wjwnrwcvlfm2atvcnnclh3x7thwrl7pgnffaw24mgws0dga4m',
+          },
         ],
       },
     }