+++ packages/composites/augur/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @chainlink/augur-adapter
 
+## 1.0.20
+
+### Patch Changes
+
+- Updated dependencies [a74101705]
+- Updated dependencies [703b60579]
+  - @chainlink/ea-bootstrap@1.9.0
+  - @chainlink/sportsdataio-adapter@1.1.18
+  - @chainlink/therundown-adapter@1.2.4
+
 ## 1.0.19
 
 ### Patch Changes