+++ packages/composites/dydx-rewards/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @chainlink/dydx-rewards-adapter
 
+## 1.0.19
+
+### Patch Changes
+
+- Updated dependencies [baa7ec876]
+- Updated dependencies [a74101705]
+- Updated dependencies [703b60579]
+  - @chainlink/ipfs-adapter@1.2.0
+  - @chainlink/ea-bootstrap@1.9.0
+
 ## 1.0.18
 
 ### Patch Changes