+++ packages/sources/finage/CHANGELOG.md
@@ -1,5 +1,14 @@
 ## v1.1.0
 
+## 1.3.5
+
+### Patch Changes
+
+- Updated dependencies [a74101705]
+- Updated dependencies [703b60579]
+  - @chainlink/ea-bootstrap@1.9.0
+  - @chainlink/ea-test-helpers@1.0.1
+
 ## 1.3.4
 
 ### Patch Changes