+++ packages/sources/oilpriceapi/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @chainlink/oilpriceapi-adapter
 
+## 2.0.4
+
+### Patch Changes
+
+- Updated dependencies [a74101705]
+- Updated dependencies [703b60579]
+  - @chainlink/ea-bootstrap@1.9.0
+  - @chainlink/ea-test-helpers@1.0.1
+
 ## 2.0.3
 
 ### Patch Changes